Add missing libX11 reference to build system
GTK on its own is not enough to ensure libX11 is properly linked with
libspice-client-gtk.  This patch adds X11_LIBS to SPICE_GTK_LIBADD_COMMON
(and X11_CFLAGS to a relevant section as well) in src/Makefile.am, and
performs an approriate pkg-config based check to determine the correct
values in configure.ac when not building for win32 or quartz.

For more info see http://bugs.gentoo.org/585118