channel-display: implement preferred video codec msgc
* SPICE_MSGC_DISPLAY_PREFERRED_VIDEO_CODEC_TYPE

This message was introduced in protocol 0.12.13 to establish client
side preference on video codec to be used in streams.

At this moment, we only introduce a new API [0] to select *the*
preferred video codec for client; In a later time, it should be
possible to use this message upon connection in order to to give
higher priority for video codecs with better performance and with
hardware decoding capabilities.

[0] spice_display_change_preferred_video_codec_type()

Note that host preference for encoding is expected to be considered
first then the client's preference.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>