streaming: Report invalid streams to the server
The error is sent using the existing client stream report message where
the dropped frame count is maxed out while the received frame count is
zero. Servers that recognize it can then switch to sending regular
screen updates for that area so the client is not stuck with a frozen
area on the screen.
This can be useful in case the client is unable to decode the stream for
some reason like a bug in the GStreamer plugins, the decoder not liking
odd video dimensions, etc.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Acked-by: Victor Toso <victortoso@redhat.com>