channel-main: demote log level from warning to debug
The volume-async happens when agent announce its capabilities. It is
common that we don't have a playback stream being played at this time
and it is even harder for a record stream.

If the audio back-end does not cache the last volume used or if this
is the first time the application is launched, this message will be
seen.

Failing to get playback or record volume async on startup is very
common making this warning too worrisome for users.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>