streaming: Allows protocol to have more codecs
The source file contains a list of fixed codecs for streaming.
This list is tested to include all codecs from spice-protocol.
If spice-protocol is extended to add more codecs this file starts
failing to compile due to missing codecs.
This patch allow to compile supporting less codecs than the
ones listed in spice-protocol.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>