gtk-session: use clear variable for array's size
By using 'num_types' variables, we have a clear variable with a clear
purpose: It will track the number of VD_AGENT_CLIPBOARD types we are
storing in types[] array.

This new variable helps:
- removing one 'for' loop which was counting the number stored types;
- doing fewer iterations in one 'for' loop

A few extra comments were included to clarify what the logic should be
doing and a extra debug was included to point out situations where the
desktop has sent us valid clipboard data but none will be sent the
guest.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>