spice_channel_read_wire: prefer while(TRUE) instead of goto
Although this is likely to be a single loop iteration based on the
implementation of g_coroutine_socket_wait(), using goto to reiterate
in the code should be avoided based on spice style recommendation.

This also make it easier to add follow up changes that can increase
the loop iterations.

This patch changes:
 * The reread label with a while(TRUE);
 * The goto keyword with continue;

All other changes are only related to new indentation.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>