gtk-session: always reply to agent's clipboard request
This fixes a regression from 7b0de6217670e0f668aff2949f

We need to reply back to the agent all clipboard requests even in case
of failure otherwise it will have a pending request. The following
error message can be seen afterwards, in the linux agent, when client
sends down some clipboard data:

 > clipboard: selection requests pending on clipboard ownership
 > change, clearing

An easy way to reproduce this is:
1-) In client, copy image from lo-draw (selection or ctrl+c)
2-) In guest, paste it to GEdit (mouse3 our ctrl+v)
3-) Move to the client
4-) Move back to the guest
5-) see error on vdagent logs

The reason for failure is that client's clipboard contains different
data type (image) then what was requested from guest's editor (text)

While at it, include extra debug message as it might be hard to
identify why clipboard did not work.

Resolves: rhbz#1409854
Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>