gtk-session: move atom's debug
We already iterate in all n_atoms and get its name with
gdk_atom_name(), let's move the debug there too;

While at it, move the generic SPICE_DEBUG of the function to the top.
It should help us identify when the function was triggered but not
used due the early returns.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>