display-gst: include capabilities for h264
As the comment states, incomplete GstCaps for h264 could trigger
errors in h264parse element, such as:

  gst_h264_parse_set_caps: video/x-h264 caps without
  codec_data or stream-format

This would make h264parse to ignore the capabilities that were set.

As spice-server is encoding as byte-stream it should be fine to set
this value here too.

Any other errors to h264 format should either be reported to
GStreamer or fixed by improving the spice-protocol.

The follow up patch will identify elements in GstRegistry based on
GstCaps so this is a necessary change to have.

This is also a preparatory patch to use Playbin element to create the
pipeline. Without this, Playbin or typefind will fail to recognize the
stream as H264.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>