display-gst: remove SPICE_GSTVIDEO_AUTO check
The intention behind SPICE_GSTVIDEO_AUTO environment variable was to
easily test the decoding elements given by GStreamer while using the
decodebin element in our pipeline.

The usage of decodebin was disabled by default as it could trigger
different issues such as the usage of unstable vaapi elements [0].

[0] See: https://bugs.freedesktop.org/show_bug.cgi?id=90884

A follow-up patch will use playbin to create the pipeline. Playbin is
very similar to decodebin but it'll provide the whole pipeline which
should bring less maintenance to spice-gtk.

Further notes:
- Vaapi elements are more stable now and it should not be a problem to
  use it.
- At this moment, there is no automatic enforcement from spice-gtk to
  make usage of any video-codecs besides mjpeg. Application would need
  to send preferred-video-codec-type message to trigger video decoding
  with GStreamer.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>