fix check for 16 bit cursor mask
The size was computed using a 32 bit image while the cursor was
just 16 bit. This caused the usage of some invalid bit mask (the
bit mask is used to understand where the cursor is transparent).
This was clearly visible as the mouse cursor was usually mostly
black with some apparently random transparent pixels.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>