usbredir: fix deadlock on error
The current code deadlocks when a USB error occurs, releasing the
channel lock before the idle is called fixes this problem.

More specifically when an error occurs, we queue a call to device_error
in an idle, and switch to the main context. The device_error() calls
spice_usbredir_channel_disconnect_device() which needs the channel to be
unlocked as it calls spice_usbredir_channel_lock().

Examples:
- Redirect USB Storage device to the guest and copy some big file. While
  copy is ongoing, unplug the USB Storage from the client machine;
- Unplug the usb device from the client machine while the USB Storage
  device is being identified in the guest.
- With the auto-redirection option enabled, plug and unplug the USB
  Storage Device frequently.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1353528
Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>