option: deprecate a few esoteric options from --spice group
I propose to deprecate the options that have very unclear results, or
that are unimplemented by drivers. They were added for historical
reasons, mostly to match with spicec options, but aren't
user-friendly. Some of them could be moved in spicy options or a new
debug/tune option group. For now, mark them as deprecated so they can
be removed in a future release.

Imho, --spice-preferred-compression is a good candidate too, although
users may be more familiar with this option, at least the 'off' case
may have a clear impact.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>