channel-display: keep one struct array for video-codecs data
On channel-display.c, for each SpiceVideoCodecType we need:
 - Its enum type;
 - The associated capability;
 - A name;

On channel-display-gst.c, for each SpiceVideoCodecType we need:
 - Associated decoding elements for the pipeline;
 - Associated GstCaps, also for the video pipeline;

Follow up patch will need the associated name for given
SpiceVideoCodecType.

Moving gst_opts[] array to channel-display-priv.h to be reused. This
should also make slightly simpler when supporting a new video codec in
the future.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>