file-xfer: Inform client of errors on init of xfer
With SpiceFileTransferTask we suggest that Spice clients watch
(signal) SpiceMainChannel::new-file-transfer so they can follow the
transfer's progress till the moment it is finished.

The signal SpiceFileTransferTask::finished informs if an error happens
to the application.

This patch solves the problem of SpiceFileTransferTask::finished not
being emitted when the agent is not connected nor when file-transfer
is disabled in the host.

We should first emit SpiceMainChannel::new-file-transfer followed up
by SpiceFileTransferTask::finished, which is done by the function
spice_file_transfer_task_completed().

As channel-main chains up the "finish" signal by removing the
SpiceFileTransferTask from its GHashTable, we have to change from
GHashTableIter to a simple GList of keys.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1373830
Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Pavel Grunt <pgrunt@redhat.com>