win-usb-dev: fix device arrival event logic
https://bugzilla.redhat.com/show_bug.cgi?id=1425961
If attached new device when one device with the same vid
and pid already present, the notification is ignored and
attached device is not redirected (if auto share set) and
not displayed in USB devices widget.

This commit reverts older commit f9631cd6f8, which was
intended to solve problem when bus:addr is sometimes changed when
using WinUSB. The same commit creates the limitation for automatic
redirection of second device with the same pid:vid.
Now the preferred backend for Spice-GTK on Windows is UsbDk.
In case users of newer WinUSB will still need backward compatible
behavior, consider backend-aware comparison procedure.

A device is now identified again by its bus.addr instead of vid:pid.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>
Message-Id: <1499232106-16448-1-git-send-email-yuri.benditovich@daynix.com>
Acked-by: Uri Lublin <uril@redhat.com>
[ Marc-André - Added commit comment from Uri's review. ]
Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>