gst-audio: Do not update mmtime without real audio channel
The fake channel has been introduced to get the audio volume by starting
the gstreamer's audio pipeline and querring its volume info (see commit
aa8d044417bbf60685f59163b874ecb4f157c3c9).

Hovewer starting the pipeline updates the mmtime as a side effect. This
may cause a (big) delay in displaying a video stream.

Because the fake channel is only needed to get the volume info, make
sure to not update the mm-time with it.

Reported-by: Christophe de Dinechin <dinechin@redhat.com>
Acked-by: Christophe de Dinechin <dinechin@redhat.com>