build-sys: Lower gettext requirement
spice-gtk currently requires gettext 0.19 which is a fairly recent
release (not available in el7.3 and older, nor in ubuntu 14.04 LTS).
They both have a gettext newer than 0.18.2, so we can require that.

Requiring 0.18 or 0.18.1 is throwing automake errors because of our use
of -Werror, so we side-step this by requiring 0.18.2.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>