channel-display: remove id parameter from helper function
Instead of passing the id parameter for destroy_display_stream() which
is only used for debug, let's store the id when creating the stream at
display_stream_create().

With the removal of Id parameter, we are using a gpointer for
display_stream struct to comply with GDestroyNotify type and use this,
for instance in g_clear_pointer() without warnings.

Benefits of this patch:
* We can drop a helper function in a follow up patch;
* Function is now a GDestroyNotify type;

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>