channel-display-gst: Prevent accumulating output queue
display_queue is queued with decoded frames ready to be displayed.
However current code can insert a timeout before displaying and
removing the queued frames. As the frames are not compressed the
usage of memory by this queue can became in some cases quite
huge (in the order of multiple gigabytes).
To prevent this do not queue all the frames but just one and count the
pending frames.
This way GStreamer will stop queuing after a while.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Snir Sheriber <ssheribe@redhat.com>