Revert "channel-usbredir: Fix crash on channel-up"
This reverts commit 291f3e4419e6fb4077ae43a5e09eb1c37b9dd729.

Follow up patch should address rhbz#1399838 mentioned in commit above
in a different way.

Major reason to revert is that the SpiceUsbDeviceManager object is
kept in SpiceSession as an easy way to share it between different
SpiceUsbredirChannel while SpiceSession itself does not use it. This
causes problems on migration as we start a new session with a new
SpiceUsbDeviceManager object while the previous one still exists.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1558043
Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>