Fix using NULL main channel reference
The main channel reference d->main of SpiceWidget is initialized on
construction. However, it is used a couple of times before that moment
while setting the default values of some properties, like resize-guest.
This results in some annoying critical log messages. This commit makes
sure that d->main is not used in such situations if it is NULL.